requirements.txt
Flask==2.0.3
python-dotenv==0.19.0
requests==2.26.0